+++ noderedpy/__init__.py
@@ -10,7 +10,7 @@ from .__property__ import (
     SpinnerProperty, CheckBoxProperty, ComboBoxProperty
 )
 
-__version__ = "0.2.9"
+__version__ = "0.2.10"
 __author__ = "oyajiDev"
 __email__ = "this.dev.somehit@gmail.com"